+++ src/internet/model/ipv4-l3-protocol.h
@@ -78,6 +78,8 @@ class Ipv4L3Protocol : public Ipv4
      * @return the object TypeId
      */
     static TypeId GetTypeId();
+
+    NS_DEPRECATED_3_49("Use the numbers defined in iana-ieee802-numbers.h")
     static constexpr uint16_t PROT_NUMBER = iana::Ieee802Numbers::IPV4; //!< Protocol number
 
     Ipv4L3Protocol();