+++ src/lte/model/lte-enb-net-device.cc
@@ -25,6 +25,7 @@
 #include "ns3/abort.h"
 #include "ns3/callback.h"
 #include "ns3/enum.h"
+#include "ns3/iana-ieee802-numbers.h"
 #include "ns3/ipv4-l3-protocol.h"
 #include "ns3/ipv6-l3-protocol.h"
 #include "ns3/llc-snap-header.h"
@@ -384,8 +385,8 @@ bool
 LteEnbNetDevice::Send(Ptr<Packet> packet, const Address& dest, uint16_t protocolNumber)
 {
     NS_LOG_FUNCTION(this << packet << dest << protocolNumber);
-    NS_ABORT_MSG_IF(protocolNumber != Ipv4L3Protocol::PROT_NUMBER &&
-                        protocolNumber != Ipv6L3Protocol::PROT_NUMBER,
+    NS_ABORT_MSG_IF(protocolNumber != iana::Ieee802Numbers::IPV4 &&
+                        protocolNumber != iana::Ieee802Numbers::IPV6,
                     "unsupported protocol " << protocolNumber
                                             << ", only IPv4 and IPv6 are supported");
     return m_rrc->SendData(packet);