+++ src/openflow/model/openflow-switch-net-device.cc
@@ -6,6 +6,7 @@
 
 #include "openflow-switch-net-device.h"
 
+#include "ns3/iana-ieee802-numbers.h"
 #include "ns3/tcp-l4-protocol.h"
 #include "ns3/udp-l4-protocol.h"
 
@@ -480,11 +481,11 @@ OpenFlowSwitchNetDevice::BufferFromPacket(Ptr<const Packet> constPacket,
     eth_header* eth_h = (eth_header*)buffer->l2;
     dst.CopyTo(eth_h->eth_dst); // Destination Mac Address
     src.CopyTo(eth_h->eth_src); // Source Mac Address
-    if (protocol == ArpL3Protocol::PROT_NUMBER)
+    if (protocol == iana::Ieee802Numbers::ARP)
     {
         eth_h->eth_type = htons(ETH_TYPE_ARP); // Ether Type
     }
-    else if (protocol == Ipv4L3Protocol::PROT_NUMBER)
+    else if (protocol == iana::Ieee802Numbers::IPV4)
     {
         eth_h->eth_type = htons(ETH_TYPE_IP); // Ether Type
     }
@@ -498,7 +499,7 @@ OpenFlowSwitchNetDevice::BufferFromPacket(Ptr<const Packet> constPacket,
 
     // We have to wrap this because PeekHeader has an assert fail if we check for an Ipv4Header that
     // isn't there.
-    if (protocol == Ipv4L3Protocol::PROT_NUMBER)
+    if (protocol == iana::Ieee802Numbers::IPV4)
     {
         Ipv4Header ip_hd;
         if (packet->PeekHeader(ip_hd))
@@ -547,7 +548,7 @@ OpenFlowSwitchNetDevice::BufferFromPacket(Ptr<const Packet> constPacket,
         }
     }
 
-    if (protocol == Ipv4L3Protocol::PROT_NUMBER)
+    if (protocol == iana::Ieee802Numbers::IPV4)
     {
         ip_header* ip_h = (ip_header*)buffer->l3;
         if (ip_h->ip_proto == TcpL4Protocol::PROT_NUMBER)